Add registry package for static DB data
- Implement package to store static database information
- Refactor planet logic to use the new registry

+++ internal/httpserver/planethandlers/mapping.go
@@ -6,7 +6,7 @@ func toTransportPlanet(p models.Planet) GetPlanetResponse {
 	buildings := make(map[string]BuildingInfo)
 	for bType, bInfo := range p.Buildings {
 		buildings[string(bType)] = BuildingInfo{
-			Level:       bInfo.Level,
+			Level:       uint8(bInfo.Level),
 			Bonuses:     bInfo.Bonuses,
 			ProductionS: bInfo.ProductionS,
 			UpdatedAt:   bInfo.UpdatedAt,
@@ -33,7 +33,7 @@ func toTransportPlanet(p models.Planet) GetPlanetResponse {
 func toTransportBuildingStats(bs models.BuildingStats) GetBuildStatsResponse {
 	return GetBuildStatsResponse{
 		Type:                 string(bs.Type),
-		Level:                bs.Level,
+		Level:                uint8(bs.Level),
 		MetalCost:            bs.MetalCost,
 		CrystalCost:          bs.CrystalCost,
 		GasCost:              bs.GasCost,
